feat: Refactor to lib
